eslint: refine rules during group discussion
removed the ban-types: e.g. string should be used over String
removed camelCase: because it was the default already
removed generic-type-naming: was set to uppercase-only /^[A-Z]+$/ which does not make sense
removed interface-name-prefix: because it was very close to the default
removed no-extraneous-class: can re-add once it is an issue
removed no-undef: TypeScript does this for us, also it was overridden for .ts files in `overrides`
removed @typescript-eslint/no-require-imports: is for node / resource loader files e.g. browser tests or weback config
removed @typescript-eslint/no-var-requires: is for node / resource loader files e.g. browser tests or weback config