Relax peer dependency version requirement
We want to use a recent version of `eslint-plugin` and `parser`,
but we don't really care about the minor version here. In particular
there is a peer dependency conflict if both are set to 8.0.1 since
`eslint-plugin` at 8.0.1 has a peer dependency on `parser` at
exactly 8.0.0.

Relax the version requirement per the suggestion from the NPM docs
(https://nodejs.org/en/blog/npm/peer-dependencies) so that we can
accept a peer dependency of any version of the packages greater
than 8.

Bug: T370126